Fix null safety in upload progress

+++ flutter_web/lib/pages/admin_downloads_page.dart
@@ -271,7 +271,7 @@ class _AdminDownloadsPageState extends State<AdminDownloadsPage> {
         onProgress: hasUpload
             ? (sent, total) {
                 if (!mounted) return;
-                final progress = total > 0 ? (sent / total).clamp(0, 1) : 0.0;
+                final progress = total > 0 ? (sent / total).clamp(0, 1).toDouble() : 0.0;
                 setState(() {
                   _uploadProgress = progress;
                   _uploadStatus = 'Upload läuft… ${(progress * 100).toStringAsFixed(0)} %';
